requirements.txt
fastapi==0.95.1
influxdb-client==1.36.1
python-dotenv==0.21.0
SQLAlchemy==2.0.2
PyMySQL==1.0.3
cryptography==40.0.2
